Dev backend orders items fix

+++ backend/src/main/java/com/nocountry/backend/repository/IOrderStatusRepository.java
@@ -4,5 +4,5 @@ import com.nocountry.backend.model.entity.OrderStatus;
 import com.nocountry.backend.model.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IOrderStatusRepository  extends JpaRepository<OrderStatus, Integer> {
+public interface IOrderStatusRepository  extends JpaRepository<OrderStatus, Long> {
 }
