+++ backend/src/main/java/com/nocountry/backend/repository/IShippingMethodRepository.java
@@ -4,5 +4,5 @@ import com.nocountry.backend.model.entity.ShippingMethod;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IShippingMethodRepository  extends JpaRepository<ShippingMethod, Integer> {
+public interface IShippingMethodRepository  extends JpaRepository<ShippingMethod,Long> {
 }
